Room_UnitTest
This file imports the Room class you made in the ex47.game module so
that you can do tests on it. There is then a set of tests that are
functions starting with test_. Inside each test case there's a bit of
code that makes a room or a set of rooms, and then makes sure the rooms
work the way you expect them to work. It tests out the basic room
features, then the paths, then tries out a whole map.